requirements.txt
Django==5.1.5
django_environ==0.12.0
environ==1.0
python-dotenv==1.1.0
whitenoise==6.4.0